pyproject.toml
[tool.poetry]
name = "basic_langchain"
version = "0.1.0"
description = "FastAPI backend using LangChain with Python 3.12"
authors = ["Suhee Kang <comeevery@gmail.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.103.0"
uvicorn = "^0.23.1"
openai = "^1.51.0"
pydantic = "^2.2.1"
python-dotenv = "^1.0.0"
langgraph = "^0.2.34"
langchain = "^0.3.1"
langchain-community = "^0.3.1"
langchain-core = "^0.3.8"
langchain-upstage = "^0.3.0"
langchain-openai = "^0.2.1"
faiss-cpu = "^1.7.4"
python-Levenshtein = "^0.26.0"
rapidfuzz = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000"
